Added Game End UI and functionality
Added a function to the gamecontroller script that when called triggers one of the endings of the game based on an Enum called Ending. Added a prefab for the EndScreen and added a UI and USS for the GameEnd UI as well as the assets the endings need to display their associated screen.